# api/docs/swagger.yaml
definitions:
  structs.CompanyStruct:
    properties:
      cover:
        type: string
      createdAt:
        description: Services  ServiceStruct `json:"serviceStruct"`
        type: string
      id:
        type: integer
      name:
        $ref: '#/definitions/structs.Language'
      slogan:
        $ref: '#/definitions/structs.Language'
      updatedAt:
        type: string
    type: object
  structs.Error:
    properties:
      message:
        type: string
    type: object
  structs.Language:
    properties:
      ki:
        type: string
      ru:
        type: string
      uz:
        type: string
    type: object
  structs.PackageStruct:
    properties:
      capacity:
        type: integer
      code:
        type: string
      cost:
        type: integer
      createdAt:
        type: string
      description:
        $ref: '#/definitions/structs.Language'
      id:
        type: integer
      name:
        $ref: '#/definitions/structs.Language'
      setId:
        type: integer
      updatedAt:
        type: string
    type: object
  structs.ServiceStruct:
    properties:
      companyId:
        type: integer
      createdAt:
        type: string
      description:
        $ref: '#/definitions/structs.Language'
      id:
        type: integer
      name:
        $ref: '#/definitions/structs.Language'
      updatedAt:
        type: string
    type: object
  structs.SetStruct:
    properties:
      createdAt:
        type: string
      description:
        $ref: '#/definitions/structs.Language'
      id:
        type: integer
      name:
        $ref: '#/definitions/structs.Language'
      serviceId:
        type: integer
      updatedAt:
        type: string
    type: object
  structs.StandardErrorModel:
    properties:
      error:
        $ref: '#/definitions/structs.Error'
    type: object
info:
  contact: {}
paths:
  /v1/companies:
    get:
      consumes:
      - application/json
      description: This API for getting list of companies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/structs.CompanyStruct'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: ListCompanies
      tags:
      - company
    post:
      consumes:
      - application/json
      description: This API for creating a new company
      parameters:
      - description: CompanyCreateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.CompanyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.CompanyStruct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: CreateCompany
      tags:
      - company
  /v1/companies/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting the company
      parameters:
      - description: CompanyId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: DeleteCompany
      tags:
      - company
    get:
      consumes:
      - application/json
      description: This API for getting company detail
      parameters:
      - description: CompanyId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.CompanyStruct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: GetCompany
      tags:
      - company
    put:
      consumes:
      - application/json
      description: This API for updating company
      parameters:
      - description: CompanyId
        in: path
        name: id
        required: true
        type: string
      - description: CompanyUpdateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.CompanyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.CompanyStruct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: UpdateCompany
      tags:
      - company
  /v1/packages:
    get:
      consumes:
      - application/json
      description: This API for getting list of packages
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/structs.PackageStruct'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: GetListPackages
      tags:
      - package
    post:
      consumes:
      - application/json
      description: This API for creating a new package
      parameters:
      - description: PackageCreateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.PackageStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.PackageStruct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: CreatePackage
      tags:
      - package
  /v1/packages/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting the package
      parameters:
      - description: PackageId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: DeletePackage
      tags:
      - package
    get:
      consumes:
      - application/json
      description: This API for getting package detail
      parameters:
      - description: PackageId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.PackageStruct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: GetPackage
      tags:
      - package
    put:
      consumes:
      - application/json
      description: This API for updating package
      parameters:
      - description: PackageId
        in: path
        name: id
        required: true
        type: string
      - description: PackageUpdateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.PackageStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.PackageStruct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: UpdateSetUpdatePackage
      tags:
      - package
  /v1/services:
    get:
      consumes:
      - application/json
      description: This API for getting list of services
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/structs.ServiceStruct'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: ListServices
      tags:
      - service
    post:
      consumes:
      - application/json
      description: This API for creating a new service
      parameters:
      - description: ServiceCreateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.ServiceStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.ServiceStruct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: CreateService
      tags:
      - service
  /v1/services/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting the service
      parameters:
      - description: ServiceId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: DeleteService
      tags:
      - service
    get:
      consumes:
      - application/json
      description: This API for getting service detail
      parameters:
      - description: ServiceId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.ServiceStruct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: GetService
      tags:
      - service
    put:
      consumes:
      - application/json
      description: This API for updating service
      parameters:
      - description: ServiceId
        in: path
        name: id
        required: true
        type: string
      - description: ServiceUpdateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.ServiceStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.ServiceStruct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: UpdateService
      tags:
      - service
  /v1/sets:
    get:
      consumes:
      - application/json
      description: This API for getting list of sets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/structs.SetStruct'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: GetListSets
      tags:
      - set
    post:
      consumes:
      - application/json
      description: This API for creating a new set
      parameters:
      - description: SetCreateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.SetStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.SetStruct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: CreateSet
      tags:
      - set
  /v1/sets/{id}:
    delete:
      consumes:
      - application/json
      description: This API for deleting the set
      parameters:
      - description: SetId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: DeleteSet
      tags:
      - set
    get:
      consumes:
      - application/json
      description: This API for getting set detail
      parameters:
      - description: SetId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.SetStruct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: GetSet
      tags:
      - set
    put:
      consumes:
      - application/json
      description: This API for updating set
      parameters:
      - description: SetId
        in: path
        name: id
        required: true
        type: string
      - description: SetUpdateRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.SetStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.SetStruct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.StandardErrorModel'
      summary: UpdateSet
      tags:
      - set
swagger: "2.0"